feat(hooks): Create useRecipeDetails hook
This commit introduces the useRecipeDetails hook, which is responsible for fetching the detailed information of a specific recipe from TheMealDB API. The hook takes a recipe ID as input and returns the recipe  data, along with loading and error states. This encapsulates the logic for fetching recipe details, making the RecipeDetailsPage component cleaner and more focused on presentation.